refactor: apply refactor in topbar component
