+++ src/components/Banner/BannerCounter/index.js
@@ -1,5 +1,5 @@
-import List from "../../Header/TopBar/List"
-import ItemList from "../../Header/TopBar/ItemList"
+import List from "../../Common/List"
+import ItemList from "../../Common/ItemList"
 import { useEffect, useState } from "react"
 
 const twoDaysInSeconds = 172800;
