+++ src/components/Common/Logo/index.js
@@ -1,0 +1,10 @@
+import Title from "../Title"
+import './styles.css'
+
+export default function Logo() {
+  return (
+    <section className="logo">
+      <Title>Electro</Title>
+    </section>
+  )
+}
