+++ src/components/Common/Title/index.js
@@ -1,0 +1,7 @@
+
+
+export default function Title({ children }) {
+  return (
+    <h1>{children}</h1>
+  )
+}
