+++ src/components/Header/index.js
@@ -1,4 +1,5 @@
 import TopBar from "./TopBar"
+import Logo from "../Common/Logo"
 import InputForm from "./InputForm"
 import FavoriteIcon from "./FavoriteIcon"
 import CartIcon from "./CartIcon"
@@ -7,18 +8,16 @@ import './styles.css'
 
 export default function Header() {
   return (
-    <section className="header">
+    <>
       <TopBar />
-      <section className="header__container">
-        <section className="header__title">
-          <h1>Electro</h1>
-        </section>
+      <header className="header">
+        <Logo />
         <InputForm />
         <section className="header__icon">
           <FavoriteIcon />
           <CartIcon />
         </section>
-      </section>
-    </section>
+      </header>
+    </>
   )
 }
